+++ resources/views/layouts/auth.blade.php
@@ -23,8 +23,8 @@
     <meta property="og:url" content="{{url()->current()}}"/>
     <meta property="og:image" content="{{config('content.image','/orchid/img/background.jpg')}}"/>
 
-    <link rel="stylesheet" href="{{asset('/orchid/css/orchid.css')}}" type="text/css"/>
-    <script async="async" src="{{asset('/orchid/js/orchid.js')}}" type="text/javascript"></script>
+    <link rel="stylesheet" href="{{mix('/css/orchid.css','orchid')}}" type="text/css"/>
+    <script async="async" src="{{mix('/js/orchid.js','orchid')}}" type="text/javascript"></script>
 </head>
 <body>
 
@@ -60,7 +60,7 @@
 
             @yield('content')
 
-            <div class="pull-bottom">
+            <div class="pull-bottom">t
                 <div class="m-b-lg clearfix v-center">
                     <div class="col-sm-3 col-md-2">
                         <a href="https://orchid.software/"><img alt="Orchid"