+++ src/Platform/Http/Controllers/Systems/MediaController.php
@@ -254,18 +254,16 @@ class MediaController extends Controller
     public function upload(Request $request)
     {
         try {
-            $path = $request->file->store($request->upload_path, $this->filesystem);
+            foreach ($request->files as $file){
+                $path = $file->move(Storage::disk($this->filesystem)->getDriver()->getAdapter()->getPathPrefix(), $file->getClientOriginalName() . '.'. $file->getClientOriginalExtension());
+            }
             $success = true;
             $message = 'Successfully uploaded new file!';
         } catch (Exception $e) {
+            $path = '';
             $success = false;
             $message = $e->getMessage();
         }
-
-        $realPath = Storage::disk($this->filesystem)->getDriver()->getAdapter()->getPathPrefix() . $path;
-        Image::make($realPath)->orientate()->save();
-
-        $path = preg_replace('/^public\//', '', $path);
 
         return response()->json(compact('success', 'message', 'path'));
     }